fix(tests): align integration suite markers
Move shared subprocess process-group setup into the test harness and mark expensive tests so CI can run the intended suite split reliably.
